examples, src: Merge ThAPI and ThFeedAPI
