+++ src/cbapi/psc/threathunter/__init__.py
@@ -7,4 +7,3 @@ from cbapi.six import PY3
 from .rest_api import CbThreatHunterAPI
 from cbapi.psc.threathunter.models import Process, Event, Tree, Feed, Report, IOCs, IOC_V2, Watchlist
 from cbapi.psc.threathunter.query import QueryBuilder
-from .feed_api import CbThreatHunterFeedAPI
