--- src/cbapi/psc/threathunter/feed_api.py
@@ -1,76 +1,0 @@
-from cbapi.connection import BaseAPI
-from cbapi.errors import ApiError
-from cbapi.psc.threathunter.models import ReportSeverity
-import logging
-log = logging.getLogger(__name__)
-
-
-class CbThreatHunterFeedAPI(BaseAPI):
-    """The main entry point into the Cb ThreatHunter PSC Feed API.
-
-    :param str profile: (optional) Use the credentials in the named profile when connecting to the Carbon Black server.
-        Uses the profile named 'default' when not specified.
-
-    Usage::
-
-    >>> from cbapi.psc.threathunter import CbThreatHunterFeedAPI
-    >>> from cbapi.psc.threathunter.models import Feed
-    >>> cb = CbThreatHunterFeedAPI(profile="production")
-    >>> for feed in cb.select(Feed):
-    >>>    print(feed.name)
-    """
-    def __init__(self, *args, **kwargs):
-        super(CbThreatHunterFeedAPI, self).__init__(product_name="psc", *args, **kwargs)
-        self._lr_scheduler = None
-
-    def _perform_query(self, cls, **kwargs):
-        return cls._query_implementation(self)
-
-    def create(self, cls, data=None):
-        """Creates a new model.
-
-        >>> feed = cb.create(Feed, feed_data)
-
-        :param cls: The model being created
-        :param data: The data to pre-populate the model with
-        :type data: dict(str, object)
-        :return: an instance of `cls`
-        """
-        return cls(self, initial_data=data)
-
-    def validate_query(self, query):
-        """Validates the given IOC query.
-
-        >>> cb.validate_query("process_name:chrome.exe") # True
-
-        :param str query: the query to validate
-        :return: whether or not the query is valid
-        :rtype: bool
-        """
-        args = {"q": query}
-        resp = self.get_object("/pscr/query/v1/validate", query_parameters=args)
-
-        return resp.get("valid", False)
-
-    def convert_query(self, query):
-        """Converts a legacy CB Response query to a ThreatHunter query.
-
-        :param str query: the query to convert
-        :return: the converted query
-        :rtype: str
-        """
-        args = {"query": query}
-        resp = self.get_object("/threathunter/feedmgr/v1/query/translate", query_parameters=args)
-
-        return resp.get("query")
-
-    @property
-    def custom_severities(self):
-        """Returns a list of active :py:class:`ReportSeverity` instances
-
-        :rtype: list[:py:class:`ReportSeverity`]
-        """
-        # TODO(ww): There's probably a better place to put this.
-        resp = self.get_object("/threathunter/watchlistmgr/v1/severity")
-        items = resp.get("results", [])
-        return [self.create(ReportSeverity, item) for item in items]
